Use secret types to select the authentication scheme
The current version of this ADR proposes to use one secret type
per protocol, which is somewhat limiting. It would be reasonable to
consider the use of a basic auth secret for authentication to an OCI
registry, for instance. Also, the use of the secret type to select the
authentication scheme would help to reduce any potential ambiguity.

In the case of OCI registries, the ADR already explicitly mentions the
dockerconfigjson secret type. For HTTP[S], it falls back on the Opaque
type and attempts to figure out the appropriate authentication method
based on the contents of the secret. This is inconsistent with the way
it works for OCI, and can be made consistent by adding a custom secret
type for the case that k0s wants to cover, but for which there's no
predefined type available upstream.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>