Fix deletion of data and kube root dirs
Make sure that we don't have anything mounted under those directories so
we don't delete persistent data.

We do this py parsing /proc/mounts in reverse order as it is listed in
mount order, and then we unmount anything that is under our directories
before we delete them.

Don't umount datadir itself if it is on a separate partition/mount

Fixes https://github.com/k0sproject/k0s/issues/4318

Signed-off-by: Natanael Copa <ncopa@mirantis.com>