Correctly detect DNS address computation errors
The existing code is unable to compute meaningful DNS addresses if
the service CIDR doesn't describe a v4 network. Error out whenever the
network IP is not a v4 address until it is clear how to properly compute
the DNS address for a v6 network. Also, copy `address` by value so that
changes to it don't change `ipnet` as well.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>