Move all kubelet config logic into the write method
This allows for dissolving the config struct and make the assignments
directly to the generated kubelet configuration struct. Removes
unnecessary indirection and concentrates more of the decision making
into the appropriate method.

Defer OS-specific decisions about the path to resolv.conf to the
specific function, so that its return values are meaningful even if it's
not called on Linux.

Set the cgroups settings and do systemd detection only on Linux, as
these are Linux-specific and don't exist on other operating systems.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>