Re-add adopters to docs, as an include
This is better than the symlink approach, because it allows the file
to live in its "canonical" place in the repository, while also avoiding
duplication when including the adopters in the docs. In fact, this opens
up more fine-grained control over how they are included, should that
become necessary in the future.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>