Overhaul container entrypoint script and Docker docs
Add some extra smarts to the k0s container entrypoint script. By adding
cgroup setup capabilities, it's no longer necessary to use the host's
cgroups when running k0s in a container. Track down and document the
necessary security-related container flags, so there's a slightly
more audited alternative to just using --privileged. Re-add the /run
directory as tmpfs, as the data in it is not meant to be persistent and
could potentially cause problems when containers are restarted.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>