Fix link to adopters in the project README
After moving the adopters file to the project root, the link in the
project README pointed to the include file for mkdocs. Fix this by
making it point to the real thing again. Since rewriting the link
doesn't really work when using it as an include, duplicate this little
paragraph in both READMES.

Fixes: 3d57e412e ("Move adopters.md back into the project root")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>