Reuse boolean for toggling the k0s endpoint reconciler
In the controller start function, the condition to determine whether to
use the k0s endpoint reconciler was written twice in very distant parts
of that function. Reuse the boolean value instead. Also, rename it to
a more appropriate name: There are actually two different reconcilers
at play: The default one built into the API server, and k0's own
implementation of it. Make it clear which implementation the boolean
refers to.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>