Use empty string to disable Docker integration
The previous sentinel value of "false" may have given the impression
that "true" is the correct value to enable Docker build integration.
Also, "false" is actually an executable provided by coreutils. An empty
string can never point to a potentially executable file, and won't give
the impression that DOCKER is some sort of boolean variable.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>