Tweak and document disabling Docker build integration
Don't default EMBEDDED_BINS_BUILDMODE to none when disabling Docker
integration. This makes it clear that the binaries can't be built
without it, at the moment. Inline the K0S_BUILD_IMAGE_FILE variable and
replace it with the more generic GO_ENV_REQUISITES. Document the new
variable in the README.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>