Reject unknown sub-commands
Cobra has the property of not reporting any unrecognized sub-commands
and simply displaying the help page for any command that doesn't have a
run function set.

Add a unit test that probes all sub-commands to properly reject unknown
sub-commands. Whenever a sub-command is added that accepts arguments and
doesn't declare them as ValidArgs, it must be added to the ignore list
at the beginning of that test.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>