Encapsulate debug flag handling
Introduce internal.DebugFlags to replace debug and verbose logging
flag handling. Move the flag definitions from the "PersistentFlagSet"
to a struct method. Remove the debug flag handling from the root
command and move it to the sub-commands. This allows for the removal
of CallPersistentPreRun, since the commands no longer rely on proper
chaining of the PersistentPreRun hooks. Remove some redundant calls to
the "PersistentFlagSet" in some sub-commands.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>