fix: prevent false positives in already running k0s instance detection
The main bugfix is to check if a possibly running pid from the runtime
config not only exists as a process, but also is the same excutable
image that is checking.

All platforms use a common library for ps-style checks now in this
implementation.

This changes the behavior for all platforms to use a common library
to perform the "is k0s running" checks on startup. The advantage is,
that the code is much more simple, the disadvantage that an additional
library is now required with additional sub-dependencies.

This commit also refactors the function checkPid into checkInstanceRunning,
since the previous implementation did not allow for more complex error
handling and would have hidden errors. More complex error handling is
now required because the implementation is more complex.

Signed-off-by: Peter Honeder <peter.honeder@unwired.at>