Provide several sub-commands only on Linux
They won't work on non-Linux systems anyways.

Use build flags to add platform-specific sub-commands instead of runtime
checks. Also add build tag to files that only make sense on specific
platforms. This includes Autopilot, as this relies on the status socket,
which is not available on Windows at the moment.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>