Remove ConfigFromReader
The function hid the fact that there was no benefit at all to providing
the input as a reader. It merely used io.ReadAll to load the data
into memory. Removing this, and exposing to callers the fact that
the only supported input is just a byte slice, makes the code more
straightforward.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>