Let API subcommand accept the runtime config via stdin
The API subcommand is not designed to run as a standalone process.
Rather, it's intended to be executed under the supervision of a k0s
controller. Therefore, the usual way of loading the configuration
is inappropriate. Instead, have this subcommand accept the runtime
configuration via stdin. This will prevent a false fallback to a
generated default configuration, as well as loading the configuration
from a possibly existing default configuration file that has nothing to
do with the one used by the supervising process.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>