Remove unused vars variable and DeepCopy
The vars variable was never read, and the DeepCopy method was used just
to initialize it. Remove both.

The original intent was probably not to list the StartupConfigPath in
the file written to disk, but since that never worked as intended and
there were no problems with it, I think it's okay to leave things as
they are.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>