Always treat an etcd backend as joinable
Controllers will always be able to connect to an etcd backend, either by
joining as a managed etcd node, or by simply connecting to an external
cluster over the network. Add some comments about whether a backend is
considered joinable or not to both the etcd and kine storage types.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>