Decouple controller flags from CLIOptions struct
This eliminates one more global variable. Make the single and
enable-worker flags an implementation detail of flag handling, and
provide an enum to the outside world instead. Draw a clear line between
controller and worker flags. Instead of accessing the entire controller
flags struct in the worker, have a bespoke interface that acts as an
integration point between a worker and its embedding controller, if any.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>