Use shared k0s build workflow in nighty network conformance runs
This reduces quite some duplicate logic. Remove the ability to test a
specific pull request, to keep things simple. It hasn't been used in a
long time. This way, it doesn't have to be integrated into the shared
build workflow, and it's still possible to run the whole workflow on
a different branch (though this has to be a branch in the k0s repo, it
can't live on a fork). Also, the network conformance tests are now part
of the standard integration test suite anyway.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>