Sleep 5 secs after airgap image bundle creation
Airgap image bundle files are sometimes truncated. This is supposedly
due to some standard input buffers in Docker that don't get flushed
properly when the container gets shut down.

Fix it in the silly way of letting the container run for a few more
seconds, so there's time to flush everything. This is a simple but
blunt solution, since writing the bundle to files inside the container
means that the Makefile has to deal with mounting and file ownership
and permissions.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit 95a6b7baef548e2620add753641a12366cdbfeee)