Extract kubelet's CA from its kubeconfig
Instead of doing it once during the bootstrapping process. This
eliminates the need for another persistent flle in k0s's data directory,
allows the use of arbitrary kubelet bootstrap kubeconfigs (as long as
they're valid), and removes a potential panic for bootstrap kubeconfigs
that don't have a cluster called "k0s".

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>