Merge autopilot worker retry loops
It's not necessary to have two successive retry loops. In fact, it's
counterproductive because indexers can only be registered with a manager
once. A second attempt will result in an index conflict error. However,
this is not a problem if the manager is created in the same retry loop.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit fb08b11c8011b25805fb6dc72abe55b1c8924de6)