Always skip controller-runtime name validation
Controller-runtime maintains a global checklist of controller names and
does not currently provide a way to unregister the controller names used
by discarded managers. The autopilot controller and worker components
accidentally share some controller names. So it's necessary to suppress
the global name check because the order in which components are started
is not fully guaranteed for k0s controller nodes running an embedded
worker.

Fixes: 79b95965a ("Bump Kubernetes to v1.31.0")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>