Replace markdownlint GitHub Action
The current action is archived and recommends the new action. The new
action has markdownlint built in, so pinning the action version also
means pinning the markdownlint version. In addition, dependabot will
automatically bump the action version in the future as well.

Also fix some lints in the pull request template on the way.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>