Ignore replacement errors for containerd executables on Windows
When containerd is restarted on Windows, the containers previously
started by the kubelet will usually continue to run. This means that
some containerd executables will still be running, too. However, unlike
on Linux, you cannot simply replace executables that are still running
on Windows.

As a workaround, k0s will ignore such errors on Windows for the time
being. Currently, this means that after a k0s update, a Windows node
must be rebooted to complete the update. This is far from an optimal,
but since Windows support is currently pretty broken anyway, it is
a workaround with minimal impact on the existing codebase, which
facilitates further testing and bug fixing.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>