Be less strict about CPLB vs. endpoint reconciler
As CPLB will work just fine with externalAddress by implicitly disabling
the endpoint reconciler, I think it makes sense to be less strict with
users about requiring them to explicitly disable the reconciler.

Change the log statement to be Info instead of Warn, and change the tone
in the docs a bit.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>