Add more complex Docker Compose example
This shows how traefik can be used as a load balancer in front of the
control plane, along with a post-start script that sets up join tokens
in a way that doesn't require etcd to be operational.

Credits go to @tdesaules for the initial idea!

Also noteworthy: I had to adjust the security flags for /dev/kmsg to
get the examples working again. I'm pretty sure this wasn't necessary
on older kernels, but I couldn't find any reference and didn't bother
downgrading to check.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>