Fix konnectivity health check in OS tests
The new konnectivity version 0.32.0 has changed the log format. Since
the OS tests relied on pod logs to determine if all konnectivity agents
were connected to all servers, this log format change broke the check.

Since the agents also provide a metrics endpoint, it is better to use
that to check if all are connected.

See: 7afaf794d ("Bump konnectivity to v0.32.0")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>