Fix backwards compat for debug flag handling
The current version of the debug flag handling rejects debug flags
on the airgap level, i.e. "k0s airgap -v" produces an error, which was
valid in 1.32.

Add back the debug flags as hidden/deprecated on that level, and add
explicit support for them on the level where they're actually used. Do
a similar change for the config and token subcommands, so that the flags
are only available for subcommands for which they make sense.

Fixes: f9773809e ("Encapsulate debug flag handling")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>