Revamp noderole inttest
The previous test was simply assuming that the taints would be present
when the labels are updated, which makes certain assumptions about the
implementaion that are irrelevant for this test case.

Use a watch to observe both label and taint changes. Wait two minutes
for unwanted taints to appear (it is always difficult to prove the
absence of something in an eventually consistent system).

Combine the noderole and noderole-no-taints tests: Let the first
controller be the one without taints, all other controllers should be
tainted. Implement the noderole-single test in the same package as the
noderole test.

Sort the checks in the Makefile alphabetically and add an env variable
containing the test target.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>