CPLB prevent IPVS routing loops
This commit enables the load balancer feature exclusively on the master
VRRP server and disables everywhere else, we rely on keepalived
notify_master and notify_backup feature so it should be as reliable as
it gets.

Unfortunately the patch can only be applied if there is exactly one VRRP
instance.

Fixes #5178

Co-authored-by: Tom Wieczorek <twz123@users.noreply.github.com>
Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>