Introduce renovate bot
Dependabot has limited scope and it cannot update things based on regex
or things that we define in Makefiles.

This is a very basic first configuration file for renovate bot which can
update etcd and kine as a proof of concept so that we decide if we want
to implement this and invest more time updating every component.

Co-authored-by: Tom Wieczorek <twz123@users.noreply.github.com>
Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>