Wait for lease pool shutdown when stopping components
The components holding a lease pool weren't waiting for the pool to
shutdown when they were stopped. They simply canceled an internal
context. This led to a race condition with the API server shutdown.
There was no guarantee that the API server was still up and running when
the lease pool tried to invalidate its lease. This left the pool in a
stale state within the cluster, causing leader transitions to take much
more time than necessary.

Add a new channel as a return value to LeasePool that callers can use
to await pool termination. Wire this up in the components that use
LeasePool.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>