Support extensions from OCI registries with self-signed CA
Allow providing a CA certificate for OCI registries in Helm commands.
Reuse the existing v1beta1.Repository type to avoid changing CRDs.

While k0s already supports custom TLS transport for classic Helm
repositories, it is lacking the same functionality for OCI registries.
This makes it impossible to install Helm extensions from OCI registries
that use private/self-signed CA for TLS.

With this change, users can add their OCI registry as a Helm repository
and provide a CA certificate path. This WILL NOT modify the repositories.yaml
file, but instead create a custom registry client for the provided OCI
registry. This client will be used with charts that reference the same
registry in the chartName.

Fixes: https://github.com/k0sproject/k0s/issues/5877

Signed-off-by: Dmitrii Shishliannikov <dshishliannikov@mirantis.com>