Fix the OCI support backport
There's two fixes bundled as one commit since without either the tests would be broken.

Remove the basic auth option for Helm OCI client. It's included in Helm 3.17 and we're currently on 3.16. Since it's NOT working on Helm itself currently, it has not been working for k0s charts either. Hence removing it does not break things more.

Bump bootloose to 0.8.0 to get the needed extraArgs and fix the OCI needed registry machine to match that.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>