Remove BinPath and let StageExecutable return the path instead
Refactor StageExecutable to return the resolved binary path. This allows
for the BinPath function to be integrated into StageExecutable, since
its sole purpose was to support PATH lookups of binaries that weren't
bundled with k0s.

As a consequence, the containerd component had to be refactored to
pick up the correct path from the appropriate call to StageExecutable.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>