Fix the nightly OS tests Slack notification
The GitHub run ID substitution was missing the dollar sign. Replace
it with an environment variable substitution, as the run ID is
automatically available as GITHUB_RUN_ID. Add the required environment
variables for the gh CLI. Without a git checkout, it won't know the
owner and repo. It needs a secret, too. Move the other two environment
variable declarations from the job into the step, as they're only used
there.

Fixes: 20f79c856 ("Use a custom Slack notification for nightly OS tests")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>