Rewrite PingPong scripts in Go
Compile the PingPong responder parts into the test executable instead
of relying on bespoke scripts that depend on an external script
interpreter. This allows for reusing code between the tests and the
responding executable, and gives finer control over the executable's
behavior. The special PingPong responder logic is triggered in when some
special command line arguments are passed to the test executable, and
will then replace its usual testing logic.

Introduce the new Pipe type that acts as an abstraction over PingPong's
pipes. Pipe, along with its companion Listener interface, is implemented
in terms of UNIX sockets, or Windows named pipes. Having these types
in place, most of the operating system specific logic disappears, and
allows for a straight forward OS agnostic implementation of both the
test code and the executable.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>