Use Go's os.Process API when cleaning up PID files
In recent Go versions, Go's process API gained the ability to open
processes using pidfds on kernels that support it. Support was added
in Linux 5.3. This addresses a possible race condition that may occur
if the kernel reuses recycled PIDs. Keeping an open file descriptor
guarantees that the PID won't be recycled. This ensures that the signal
is sent to the intended process and not to another process that was
assigned the same recycled PID. On older kernels, it simply falls back
to the old, racy implementation that uses PIDs.

Rename newProcHandle to openPID to indicate that the handle needs to be
closed now.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>