Verify airgap bundle integrity in bundler script
Airgap image bundle files are sometimes truncated. A previous assessment
attributed this to stdout buffers in Docker not being flushed properly
when the container shuts down. The initial fix was to add a 5-second
delay after `ctr` completes. However, this appears to be the wrong
solution, or at least an incomplete one. Even with a 30-second delay,
the bundle can still be truncated.

Add an image bundle verification step directly in the bundler script
itself, so that the resulting files inside the Docker container and on
the GitHub runner can be compared.

Surprisingly, this check appears to resolve the truncation issue as a
side effect. I have no idea why that is.

See: 95a6b7bae ("Sleep 5 secs after airgap image bundle creation")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>