Implement k0s feature gates
Introduces a feature gates system to k0s, inspired by Kubernetes feature
gates, enabling controlled rollout of experimental and new features.

These are heavily inspired but kubernetes' feature gates, but we don't
need something that sophisticated and hence we're doing a much simpler
implementation for them.

Unlike in kubernetes, we've decided not to implement AllAlpha and
AllBeta at least for now.

Implements #6084

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>