Kill supervised process after stop timeout
Previously, the shutdown code looped endlessly until the child process
finished, requesting graceful termination over and over again. Change
this to a request-shutdown -> wait -> kill -> wait logic. This is
to ensure that k0s won't hang when the supervised processes can't be
terminated for whichever reason: the code will terminate, at least after
all the timeouts expired.

Use a buffered channel for the wait result, so that the goroutine
will be able to exit, even if nothing reads from the channel anymore.
Introduce fine-grained error reporting to differentiate shutdown
outcomes (graceful shutdown, forced kill, failure, and so on).

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>