Replace all \s with [ \t] in Renovate regexes
The \s character class matches _all_ whitespace, including line
terminators. That's not what we want. Use a simple match on spaces
and tabs instead. This won't include all the other Unicode whitespace
characters, but we're dealing with source code, in the end, so this
is fine. Also properly escape \n and only allow single spaces in the
comments. This makes the regex more complex and reflects what's actually
used.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>