Don't export inttest environment variables manually
Instead of manually exporting the required environment variables in the
Go test invocation, export the Make variables directly. This makes the
command line easier to read. As an exception to this rule, introduce the
new variable K0S_INTTEST_TARGET via the command line to make it clear
which target is being executed. This will allow for the removal of some
extra environment variables from the Makefile in the future because
the tests can simply use the value of that variable to determine their
configuration.

Deduce K0S_PATH from the current directory if it is not set. This
eliminates the need to define K0S_PATH in every case and relies on the
fact that the inttests are always executed with the working directory
set to their source directory.

Move the K0S_UPDATE_TO_VERSION definition to the ap-ha3x3 inttest, which
is the only one that needs it.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit fb3e54ecb97e1267f97077794a8e274fb8f397b3)