Add unit test for konnectivity versions
There has been an accidental downgrade of konnectivity in the go.mod
file that caused the k0s executable to use a different version than
the embedded konnectivity-server executable. Add a unit test similar to
the others which ensures that the go.mod and the embedded binaries use
matching versions of konnectivity.

See: 41e98fb88 ("Bump kubernetes to v1.34.0-beta.0")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>