Enable concurrency for k0s airgap bundle-artifacts
Adds a flag to define the concurrency level and defaults to 3. The
defualt value is 3 because the default value of ORAS is 3 and because
after some primitive testing in my environment it seems pretty optimal:

1- Increasing the value from 1 to 2 is a massive difference. Time is
reduced consistently between 60 and 80% in my environment (NVMe, wired
600mbps optical fiber connection)
optical firewall

2- Increasing from 2 to 3 is usually around 5-20% better, but it's not
nearly as consistent

3- The difference after 3 is far too small and inconsistent to be able
to tell which value is optimal without far more exhaustive and rigorous
testing.

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>