Add "idempotency" to image repository override
When we use dynamic config, the given config is unmarshaled twice: once when we read the config from disk, and then store it in k8s api, and then second time when we read it from the API.
Since the override is wired in the unmarshaling it'll be invoked twice for a given image. Now the second time it'll be invoked for an image that is already overridden. This results in duplicating the possible path element(s) in the override repo.

To fix this I added prefix matching to prevent overriding an image which already has the given repo prefix.

Fixes #6230

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>