Allow apiVersion and kind in kubelet config validation
Let users specify the apiVersion and kind for KubeletConfigurations in
worker profiles, so that they don't need to be stripped out when copying
otherwise valid KubeletConfigurations into worker profiles.

Fail fast and skip field validations if JSON decoding fails, since those
validations would be irrelevant or misleading in that case.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>