Make the image download concurrency a field of the bundler
As this is a "configuration" flag, this should go to the struct, instead
of being passed as an argument to the Run method.

Also add some flag validation: What does a concurrency of -42 mean? ORAS
will just fallback to its default value in that case. As this is a bit
weird to document on the CLI flag, and given that k0s already hard-codes
a default value by itself, it makes sense to disallow non-positive
concurrency values completely.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>