Add cgroup auto-detection for kubelet & containerd
This PR introduces cgroups auto-detection to select the driver (systemd/cgroupsfs) based on the operating environment.
The auto-detection respects an existing configuration on the node, so which ever driver was used will be still used after upgrade.
This ensures we're not gonna change the driver "in-flight" as that would probably break containers and pods.

The scope has been solely the driver detection, so what ever hard-coded, and non-optimal, cgroups have been used will still be used by default.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>