Cleanly distinguish Supervisor Cmd.Wait errors
The os.Process API is strange in that it returns an error instead of
(ProcessState, error). This makes it difficult to distinguish between
"regular" process errors and failures that occur while actually waiting
on the process.

Nevertheless, try to distinguish between these two cases to produce
more accurate log messages: If the error is nil or unwraps into
an *exec.ExitErr, treat it as a "regular" process error. Consider
everything else an error indicating a problem with waiting.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>