Don't kill processes when cleaning up PID files
It's better to leave the decision what to do about a lingering process
to the users. Try to terminate the process gracefully once and poll for
process termination. If the process didn't terminate in time, bail out
with an error.

Re-implement the Supervisor proc handle without the stdlib. The os/exec
API is not a good fit for implementing the PID file cleanup. Instead,
use a combination of dirfds to /proc/<pid> directories, along with
pidfd_send_signal (if available) to implement a race-free process
termination.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>