Allow to configure CPLB interfaces with mac addr
Introduces a new field interfaceMACAddresses which allows to define a
slice of mac addresses which k0s will use to try to find a network
interface for the VRRPInstance to use.

Also introduces another new field interfaceType so that the user can
chose between using the previously existing interface and the new
interfaceMACAddresses.

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>