Strip namespaces from cluster resources
If there is a cluster-wide resource with a namespace set , the stack
applier gets confused by it. Although the API server ignores the
namespace, the applier takes it into account for resource pruning. This
effectively resulted in the resource being applied and deleted again.

First, remove the namespace for cluster-wide resources. Then, prepare
the resources for application. This ensures that the incorrect namespace
won't end up in the cache key or the last applied configuration.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>