Simplify IPv6 image list generation
Concatenate the names directly in the Makefile. This saves a lot of
complexity in the gen-test-images-list program.

Add some missing requisites to the IPv6 test image bundle targets.

Since the container images are kept in the Makefile itself, Renovate
needs to have an eye on it, too. Add the appropriate glob patterns to
its config.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>