Return early after starting the embedded worker
When a k0s controller runs the embedded worker, the worker itself blocks
until the context is canceled. In that case, the controller doesn't need
to block again. To make this clear, return early from the controller
start method when the embedded worker start method returns.

Harmonize the log statements in the controller and worker and remove the
performance logging that would only occur upon termination.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>