Use GitHub managed runners for linux-arm64
Now that GitHub supports arm64 for Linux, there's no longer the need to
use self-hosted runners for this. This allows for running arm64 builds
and tests on forks.

Also shorten the run names to match the make targets they're running.
The old naming scheme was truncated when displayed on the GitHub
webpage.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>