Simplify Supervisor process termination helper
Don't try to reattach to the parent console after sending the Ctrl+Break
event. The only reason this was done was the assumption that it would be
required for capturing stderr in k0s. Turns out that this is not true.
Since k0s installs pipes as the child process's stdout/stderr, they
remain valid even after detaching from the console.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>