Dissolve Windows stack
Instead of deploying the necessary manifests for Windows workers in one
stack, incorporate them into the relevant stacks: calico and kubeproxy.
This allows for reusing the template values of each stack without
duplicating the logic. The old Windows stack component is now solely
responsible for counting Windows nodes. The count is broadcasted to
the kubeproxy and calico components, which update the manifests on disk
if necessary.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>