Add basic Windows smoketest
The tests setups a WSL Linux on Windows runner which allow us to run the CP (with worker enabled) on the Linux node.
The Windows host is used as a worker node.

There's bunch of preliminary steps in order to configure networking, WSL and other stuff so that the test would be as stable as possible.
The current challenge is that when Calico boots up on the Windows side, it'll create the overlay network which in turn creates bit of a flip on the general networking.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>