Omit calico windows installer host prefic for CNI_NET_DIR
Some of the installer paths use the `/host` prefix and some don't.
In this case the path is used in the _actual_ CNI config file dropped into the host
so it needs to be WITHOUT the prefix.

Fixes: 98727e458 ("Revamp Calico manifests for Windows")

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>
(cherry picked from commit 28f57ea81eb513b320af1c7f460d26a0a71adf83)